global: finish receipts backend features with global tweaks

+++ acme/app/src/main/java/org/feup/cp/acme/ui/tabs/products/ProductsAdapter.kt
@@ -49,7 +49,7 @@ class ProductsAdapter(
             dataSet[position].price.toString().plus("$")
         button.visibility = Button.VISIBLE
         button.setOnClickListener { view ->
-            val dialog = AddCartDialog(dataSet[position].id)
+            val dialog = AddCartDialog(dataSet[position].id.toInt())
             dialog.show(activity.supportFragmentManager, "addCart")
         }
     }
